pyproject.toml
[project]
name = "dummy-project"
version = "0.1.0"
description = "A multi-server AI agent framework using LangChain and MCP (Model Context Protocol) for executing math operations and weather queries through an intelligent agent interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "mcp>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
